requirements.txt
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==3.2.5
django-extensions==3.1.1
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
psycopg2==2.8.6
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
uWSGI==2.0.19.1
